Logic for a gun to shoot bullets

+++ Assets/_Scripts/Health/IDamageable.cs
@@ -1,8 +1,8 @@
 
 public interface IDamageable
 {
-    public int MaxHP { get; }
-    public int HP { get; }
+    public float MaxHP { get; }
+    public float HP { get; }
 
     public bool TakeDamage(float damage, IDamageEffect damageEffect);
 }
